--- baebae-BE/src/main/java/com/web/baebaeBE/application/answer/AnswerApplication.java
@@ -1,69 +1,0 @@
-package com.web.baebaeBE.application.answer;
-
-import com.web.baebaeBE.domain23.answer.exception.AnswerError;
-import com.web.baebaeBE.domain23.answer.service.AnswerService;
-import com.web.baebaeBE.domain.login.exception.LoginException;
-import com.web.baebaeBE.global.error.exception.BusinessException;
-import com.web.baebaeBE.infra.answer.entity.Answer;
-import com.web.baebaeBE.infra.answer.repository.AnswerMapper;
-import com.web.baebaeBE.domain.member.entity.Member;
-import com.web.baebaeBE.domain.member.repository.MemberRepository;
-import com.web.baebaeBE.infra.question.entity.Question;
-import com.web.baebaeBE.infra.question.repository.QuestionRepository;
-import com.web.baebaeBE.presentation.answer.dto.AnswerCreateRequest;
-import com.web.baebaeBE.presentation.answer.dto.AnswerDetailResponse;
-import com.web.baebaeBE.presentation.answer.dto.AnswerResponse;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.util.List;
-import java.util.stream.Collectors;
-
-@Service
-@RequiredArgsConstructor
-public class AnswerApplication {
-    private final AnswerService answerService;
-    private final AnswerMapper answerMapper;
-    private final MemberRepository memberRepository;
-    private final QuestionRepository questionRepository;
-
-    public AnswerDetailResponse createAnswer(AnswerCreateRequest request, Long memberId, MultipartFile imageFiles) {
-        Member member = memberRepository.findById(memberId)
-                .orElseThrow(() -> new BusinessException(LoginException.NOT_EXIST_MEMBER));
-        System.out.println(request.getQuestionId());
-        Question question = questionRepository.findById(request.getQuestionId())
-                .orElseThrow(() -> new BusinessException(AnswerError.NO_EXIST_QUESTION));
-
-        Answer answerEntity = answerMapper.toEntity(request, question, member);
-        Answer savedAnswerEntity = answerService.createAnswer(answerEntity, imageFiles);
-        return answerMapper.toDomain(savedAnswerEntity, "나중에 수정 요망");
-    }
-
-    public List<AnswerResponse> getAnswersByMemberId(Long memberId) {
-        List<Answer> answers = answerService.getAnswersByMemberId(memberId);
-        return answers.stream()
-                .map(AnswerResponse::of)
-                .collect(Collectors.toList());
-    }
-
-    public Page<AnswerDetailResponse> getAllAnswers(Long memberId, Pageable pageable) {
-        Page<Answer> answerPage = answerService.getAllAnswers(memberId, pageable);
-        return answerPage.map(answer -> answerMapper.toDomain(answer, "나중에 수정 요망"));
-    }
-
-    public AnswerDetailResponse updateAnswer(Long answerId, AnswerCreateRequest request, MultipartFile imageFiles) {
-        Answer updatedAnswer = answerService.updateAnswer(answerId, request, imageFiles);
-        return answerMapper.toDomain(updatedAnswer, "나중에 수정 요망");
-    }
-
-    public void deleteAnswer(Long answerId) {
-        answerService.deleteAnswer(answerId);
-    }
-
-    public void updateReactionCounts(Long answerId, int heartCount, int curiousCount, int sadCount) {
-        answerService.updateReactionCounts(answerId, heartCount, curiousCount, sadCount);
-    }
-}
