+++ baebae-BE/src/main/java/com/web/baebaeBE/domain/answer/repository/AnswerJpaRepository.java
@@ -1,0 +1,14 @@
+package com.web.baebaeBE.domain.answer.repository;
+
+import com.web.baebaeBE.domain.answer.entity.Answer;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.List;
+
+public interface AnswerJpaRepository extends JpaRepository<Answer, Long> {
+    Page<Answer> findAllByMemberId(Long memberId, Pageable pageable);
+
+    List<Answer> findByMemberId(Long memberId);
+}
