+++ baebae-BE/src/main/java/com/web/baebaeBE/domain/answer/repository/AnswerRepository.java
@@ -1,0 +1,16 @@
+package com.web.baebaeBE.domain.answer.repository;
+
+import com.web.baebaeBE.domain.answer.entity.Answer;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
+
+import java.util.List;
+import java.util.Optional;
+
+public interface AnswerRepository {
+    Optional<Answer> findByAnswerId(Long answerId);
+    List<Answer> findByMemberId(Long memberId);
+    Answer save(Answer answer);
+    Page<Answer> findAllByMemberId(Long memberId, Pageable pageable);
+    void delete(Answer answer);
+}
