--- baebae-BE/src/main/java/com/web/baebaeBE/domain23/answer/service/AnswerService.java
@@ -1,170 +1,0 @@
-package com.web.baebaeBE.domain23.answer.service;
-
-import com.web.baebaeBE.domain23.answer.exception.AnswerError;
-import com.web.baebaeBE.global.error.exception.BusinessException;
-import com.web.baebaeBE.global.firebase.FirebaseNotificationService;
-import com.web.baebaeBE.global.image.s3.S3ImageStorageService;
-import com.web.baebaeBE.infra.answer.entity.Answer;
-import com.web.baebaeBE.infra.answer.repository.AnswerRepository;
-import com.web.baebaeBE.infra.question.entity.Question;
-import com.web.baebaeBE.presentation.answer.dto.AnswerCreateRequest;
-import jakarta.transaction.Transactional;
-import lombok.RequiredArgsConstructor;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.util.ArrayList;
-import java.util.List;
-
-@RequiredArgsConstructor
-@Service
-public class AnswerService {
-    private final AnswerRepository answerRepository;
-    private final FirebaseNotificationService firebaseNotificationService;
-    private final S3ImageStorageService s3ImageStorageService;
-
-    @Transactional
-    public Answer createAnswer(Answer answer, MultipartFile imageFile) {
-        List<String> imageUrls = new ArrayList<>();
-        String musicAudioUrl = null;
-        Answer savedAnswer = answerRepository.save(answer);
-
-        try {
-            // 이미지 파일 업로드
-            if (!imageFile.isEmpty()) {
-                InputStream inputStream = imageFile.getInputStream();
-                String imageUrl = s3ImageStorageService.uploadFile(answer.getMember().getId().toString(), answer.getId().toString(), "image", 0, inputStream, imageFile.getSize(), imageFile.getContentType());
-                imageUrls.add(imageUrl);
-            }
-
-            // 파일 URL을 Answer 엔티티에 저장
-            answer.setImageFiles(imageUrls);
-            answer.setMusicAudio(answer.getMusicAudio());
-
-        } catch (Exception e) {
-            // 롤백 이후 파일 삭제 로직
-            imageUrls.forEach(url -> s3ImageStorageService.deleteFileByUrl(url)); // 이미지 파일 삭제
-            if (musicAudioUrl != null) {
-                s3ImageStorageService.deleteFileByUrl(musicAudioUrl); // 오디오 파일 삭제
-            }
-            throw new RuntimeException("Failed to create answer", e);
-        }
-        // 새로운 답변 달리면 알림 생성
-        firebaseNotificationService.notifyNewAnswer(answer.getMember(), answer);
-        // 질문의 isAnswered 상태를 true로 업데이트
-        Question question = answer.getQuestion();
-        question.setAnswered(true);
-        return savedAnswer;
-    }
-
-    @Transactional
-    public Answer updateAnswer(Long answerId, AnswerCreateRequest request, MultipartFile imageFile) {
-        Answer answer = answerRepository.findByAnswerId(answerId)
-                .orElseThrow(() -> new BusinessException(AnswerError.NO_EXIST_ANSWER));
-
-        List<String> urls = new ArrayList<>();
-        try {
-            String memberId = answer.getMember().getId().toString();
-            String fileType = "image";
-
-            // 이미지 파일 처리
-            String fileName = "image_0.jpg";
-            InputStream inputStream = imageFile.getInputStream();
-            long size = imageFile.getSize();
-            String contentType = imageFile.getContentType();
-            String url = s3ImageStorageService.uploadFile(memberId, answerId.toString(), fileType, 0, inputStream, size, contentType);
-            urls.add(url);
-
-        } catch (IOException e) {
-            throw new BusinessException(AnswerError.IMAGE_PROCESSING_ERROR);
-        }
-
-        // 이미지 URL 리스트 업데이트
-        answer.setImageFiles(urls);
-
-        // 다른 필드 업데이트
-        answer.setContent(request.getContent());
-        answer.setLinkAttachments(request.getLinkAttachments());
-        answer.setMusicName(request.getMusicName());
-        answer.setMusicSinger(request.getMusicSinger());
-
-        // 데이터베이스 저장
-        return answerRepository.save(answer);
-    }
-
-    public List<Answer> getAnswersByMemberId(Long memberId) {
-        return answerRepository.findByMemberId(memberId);
-    }
-
-    @Transactional
-    public Page<Answer> getAllAnswers(Long memberId, Pageable pageable) {
-        return answerRepository.findAllByMemberId(memberId, pageable);
-    }
-
-    @Transactional
-    public void deleteAnswer(Long answerId) {
-        Answer answer = answerRepository.findByAnswerId(answerId)
-                .orElseThrow(() -> new IllegalArgumentException("No answer found with id " + answerId));
-
-        // S3 이미지 파일 삭제
-        // https://kr.object.ncloudstorage.com/baebae-bucket/1/1/image_1.jpg
-        answer.getImageFiles().forEach(url -> {
-            String[] parts = url.split("/");
-
-            if (parts.length > 1) {
-                String fileKey = parts[parts.length - 1];
-                String[] keyParts = fileKey.split("_");
-
-                if (keyParts.length > 1) {
-                    int index = Integer.parseInt(keyParts[1].replace(".jpg", ""));
-                    String memberId = parts[parts.length - 3];
-                    String fileType = keyParts[0];
-                    s3ImageStorageService.deleteFile(memberId, answerId.toString(), fileType, index);
-                }
-            }
-        });
-
-        //S3 오디오 파일 삭제
-        // https://kr.object.ncloudstorage.com/baebae-bucket/1/2/audio.mp3
-        String url = answer.getMusicAudio();
-        String[] parts = url.split("/");
-
-        if (parts.length > 1) {
-            String[] file = parts[parts.length - 1].split("\\.");
-            String fileType = file[0];
-            String memberId = parts[parts.length - 3];
-
-            s3ImageStorageService.deleteFile(memberId, answerId.toString(), fileType, 0);
-        }
-
-        // answer 객체 삭제
-        answerRepository.delete(answer);
-    }
-
-    @Transactional
-    public void updateReactionCounts(Long answerId, int heartCount, int curiousCount, int sadCount) {
-        Answer answer = answerRepository.findByAnswerId(answerId)
-                .orElseThrow(() -> new BusinessException(AnswerError.NO_EXIST_ANSWER));
-
-        // 반응수 올라갈때마다 알림 전송
-        if (answer.getHeartCount() != heartCount) {
-            firebaseNotificationService.notifyReaction(answer.getMember(), answer.getContent(), "heart");
-        }
-        if (answer.getCuriousCount() != curiousCount) {
-            firebaseNotificationService.notifyReaction(answer.getMember(), answer.getContent(), "curious");
-        }
-        if (answer.getSadCount() != sadCount) {
-            firebaseNotificationService.notifyReaction(answer.getMember(), answer.getContent(), "sad");
-        }
-
-        // 반응 수 업데이트
-        answer.setHeartCount(heartCount);
-        answer.setCuriousCount(curiousCount);
-        answer.setSadCount(sadCount);
-        answerRepository.save(answer);
-    }
-}
