+++ baebae-BE/src/main/java/com/web/baebaeBE/global/firebase/FirebaseNotificationService.java
@@ -1,8 +1,8 @@
 package com.web.baebaeBE.global.firebase;
 
 import com.web.baebaeBE.domain.member.entity.Member;
-import com.web.baebaeBE.infra.answer.entity.Answer;
-import com.web.baebaeBE.infra.question.entity.Question;
+import com.web.baebaeBE.domain.answer.entity.Answer;
+import com.web.baebaeBE.domain.question.entity.Question;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
