--- baebae-BE/src/main/java/com/web/baebaeBE/infra/answer/repository/AnswerJpaRepository.java
@@ -1,14 +1,0 @@
-package com.web.baebaeBE.infra.answer.repository;
-
-import com.web.baebaeBE.infra.answer.entity.Answer;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface AnswerJpaRepository extends JpaRepository<Answer, Long> {
-    Page<Answer> findAllByMemberId(Long memberId, Pageable pageable);
-
-    List<Answer> findByMemberId(Long memberId);
-}
