--- baebae-BE/src/main/java/com/web/baebaeBE/infra/answer/repository/AnswerRepository.java
@@ -1,16 +1,0 @@
-package com.web.baebaeBE.infra.answer.repository;
-
-import com.web.baebaeBE.infra.answer.entity.Answer;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface AnswerRepository {
-    Optional<Answer> findByAnswerId(Long answerId);
-    List<Answer> findByMemberId(Long memberId);
-    Answer save(Answer answer);
-    Page<Answer> findAllByMemberId(Long memberId, Pageable pageable);
-    void delete(Answer answer);
-}
