+++ baebae-BE/src/main/resources/application.yml
@@ -5,3 +5,5 @@ spring:
   profiles:
     active: local
 
+server:
+  url: http://localhost:8080
